requirements.txt
Django>=4.0
djangorestframework
openai
psycopg2